+++ pyppeteer/connection.py
@@ -7,7 +7,7 @@ import asyncio
 import json
 import logging
 from typing import Awaitable, Callable, TYPE_CHECKING
-
+import concurrent.futures
 from pyee import EventEmitter
 import websockets
 
@@ -57,14 +57,27 @@ class Connection(EventEmitter):
                 except (websockets.ConnectionClosed, ConnectionResetError):
                     logger.info('connection closed')
                     break
+                await asyncio.sleep(0)
+        if self._connected:
+            asyncio.ensure_future(self.dispose())
 
-    async def _async_send(self, msg: str) -> None:
+    async def _async_send(self, msg: str, callback_id: int) -> None:
         while not self._connected:
             await asyncio.sleep(self._delay)
-        await self.connection.send(msg)
+        try:
+            await self.connection.send(msg)
+        except websockets.ConnectionClosed:
+            logger.error('connection unexpectedly closed')
+            callback = self._callbacks.get(callback_id, None)
+            if callback and not callback.done():
+                callback.set_result(None)
+                await self.dispose()
 
     def send(self, method: str, params: dict = None) -> Awaitable:
         """Send message via the connection."""
+        # Detect connection availability from the second transmission
+        if self._lastId and not self._connected:
+            raise ConnectionError('Connection is closed')
         if params is None:
             params = dict()
         self._lastId += 1
@@ -74,8 +87,8 @@ class Connection(EventEmitter):
             method=method,
             params=params,
         ))
-        logger.debug(f'SEND▶: {msg}')
-        asyncio.ensure_future(self._async_send(msg))
+        logger.debug(f'SEND: {msg}')
+        asyncio.ensure_future(self._async_send(msg, _id))
         callback = asyncio.get_event_loop().create_future()
         self._callbacks[_id] = callback
         callback.method = method  # type: ignore
@@ -111,7 +124,7 @@ class Connection(EventEmitter):
         self._closeCallback = callback
 
     def _on_message(self, message: str) -> None:
-        logger.debug(f'◀RECV: {message}')
+        logger.debug(f'RECV: {message}')
         msg = json.loads(message)
         if msg.get('id') in self._callbacks:
             self._on_response(msg)
@@ -132,9 +145,9 @@ class Connection(EventEmitter):
         self._sessions.clear()
 
         # close connection
+        if hasattr(self, 'connection'):  # may not have connection
+            await self.connection.close()
         if not self._recv_fut.done():
-            if hasattr(self, 'connection'):  # may not have connection
-                await self.connection.close()
             self._recv_fut.cancel()
 
     async def dispose(self) -> None:
@@ -190,13 +203,15 @@ class CDPSession(EventEmitter):
         callback = asyncio.get_event_loop().create_future()
         self._callbacks[_id] = callback
         callback.method: str = method  # type: ignore
-
         if not self._connection:
             raise NetworkError('Connection closed.')
-        await self._connection.send('Target.sendMessageToTarget', {
-            'sessionId': self._sessionId,
-            'message': msg,
-        })
+        try:
+            await self._connection.send('Target.sendMessageToTarget', {
+                'sessionId': self._sessionId,
+                'message': msg,
+            })
+        except concurrent.futures.CancelledError:
+            raise NetworkError("connection unexpectedly closed")
         return await callback
 
     def _on_message(self, msg: str) -> None: